Existing tests passing, old list page removed
