+++ application/strava_api/views.py
@@ -100,109 +100,6 @@ def handle_code():
   return redirect(url_for('strava_api.manage'))
 
 
-# @strava_api.route('/activities', methods=['GET', 'POST'])
-# @login_required
-def display_activity_list():
-  """Display list of strava activities to view in individual Dashboards."""
-  strava_account = StravaAccount.query.get(request.args.get('id'))
-
-  if strava_account is None:
-    return redirect(url_for('strava_api.authorize'))
-    # '?after="/strava/activities"'
-
-  client = strava_account.client
-
-  limit = int(request.args.get('limit', 25))
-  page = int(request.args.get('page', 1))
-
-  activities = client.get_activities(limit=limit)
-  activities.per_page = limit
-  activities._page = page
-  activity_list = list(activities)
-  saved_activity_id_list = [a.strava_id for a in strava_account.activities.all()]
-
-  for activity in activity_list:
-    # activity._saved = any([a.strava_id == activity.id for a in strava_account.activities])
-    activity._saved = activity.id in saved_activity_id_list
-
-  form = BatchForm()
-  if form.validate_on_submit():
-      # flash('Activities will be added in the background.')
-      # take a moment
-      
-      # Desired: do the following in the background! 
-      # And redirect immediately w/ message.
-
-      # Get all activity ids (assumes everything is a valid run rn)
-      activity_ids = [activity.id for activity in client.get_activities()]
-      print(f'Num. of activities retrieved: {len(activity_ids)}')
-
-      # Get data for each activity and create an entry in db.
-      activities = []
-      for activity_id in activity_ids:
-        df = converters.from_strava_streams(client.get_activity_streams(
-          activity_id,
-          types=['time', 'latlng', 'distance', 'altitude', 'velocity_smooth',
-             'heartrate', 'cadence', 'watts', 'temp', 'moving',
-             'grade_smooth']
-        ))
-        calc_power(df)
-
-        if 'NGP' in df.columns:
-          # Resample the NGP stream at 1 sec intervals
-          # TODO: Figure out how/where to make this repeatable.
-          # 1sec even samples make the math so much easier.
-          interp_fn = interp1d(df['time'], df['NGP'], kind='linear')
-          ngp_1sec = interp_fn([i for i in range(df['time'].max())])
-
-          # Apply a 30-sec rolling average.
-          window = 30
-          ngp_rolling = pd.Series(ngp_1sec).rolling(window).mean()          
-          ngp_ms = putil.lactate_norm(ngp_rolling[29:])
-          cp_ms = 1609.34 / (6 * 60 + 30)  # 6:30 mile
-          intensity_factor = ngp_ms / cp_ms
-          total_hours = (df['time'].iloc[-1] - df['time'].iloc[0]) / 3600
-          tss = 100.0 * total_hours * intensity_factor ** 2
-        else:
-          intensity_factor = None
-          tss = None
-
-        activity_data = client.get_activity(activity_id).to_dict()
-
-        activities.append(Activity(
-          title=activity_data['name'],
-          description=activity_data['description'],
-          created=datetime.datetime.utcnow(),  
-          recorded=dateutil.parser.isoparse(activity_data['start_date']),
-          tz_local=activity_data['timezone'],
-          moving_time_s=activity_data['moving_time'],
-          elapsed_time_s=activity_data['elapsed_time'],
-          # Fields below here not required
-          strava_id=activity_data['id'],
-          distance_m=activity_data['distance'],
-          elevation_m=activity_data['total_elevation_gain'],
-          intensity_factor=intensity_factor,
-          tss=tss,
-        ))
-
-      try:
-        db.session.add_all(activities)
-        db.session.commit()
-
-      except IntegrityError as e:
-        print('There was an error saving the activities.')
-        print(e)
-
-      return redirect('/')
-
-  return render_template(
-    'strava_api/activity_list.html',
-    activities=activity_list,
-    last_page=(len(activity_list) != limit),
-    form=form
-  )
-
-
 @strava_api.route('/manage')
 @login_required
 def manage():
