--- application/templates/strava_api/activity_list.html
@@ -1,37 +1,0 @@
-{% extends 'admin_base.html' %}
-  
-{% block header %}
-  <h2>{% block title %}Your Strava Activities{% endblock %}</h2>
-{% endblock %}
-
-{% block content %}
-  <form action="" method="POST">
-      {{ form.csrf_token }}
-      <p>{{ form.submit() }}</p>
-  </form>
-
-  {% for act in activities %}
-    <a href="/strava/{{ act.id }}?id={{ request.args.get('id') }}" {% if act._saved %}class="alert-danger"{% endif %}>{{ act.name }}</a> ({{ act.start_date_local }}) ({{ (act.distance.to('mile')).magnitude | round (1) }} mi, {{ act.total_elevation_gain.to('foot').magnitude | round(0) | int }}')
-    <br>
-  {% endfor %}
-  {% with p=request.args.get('page')|int or 1 %}
-  {% with limit=request.args.get('limit') %}
-  {% with id=request.args.get('id') %}
-    <nav aria-label="...">
-      <ul class="pagination">
-        {% if p==1 %}
-          <li class="page-item disabled"><a class="page-link" href="#"><span aria-hidden="true">&larr;</span>Newer</a></li>
-        {% else %}
-          <li class="page-item"><a class="page-link" href="/strava/activities?id={{ id }}&page={{ p - 1 }}{% if limit %}&limit={{ limit }}{% endif %}"><span aria-hidden="true">&larr;</span>Newer</a></li>
-        {% endif %}
-        {% if last_page %}
-          <li class="page-item disabled"><a class="page-link" href="#"><span aria-hidden="true">Older<span aria-hidden="true">&rarr;</span></a></li>
-        {% else %}
-          <li class="page-item"><a class="page-link" href="/strava/activities?id={{ id }}&page={{ p + 1 }}{% if limit %}&limit={{ limit }}{% endif %}">Older<span aria-hidden="true">&rarr;</span></a></li>
-        {% endif %}
-      </ul>
-    </nav>
-  {% endwith %}
-  {% endwith %}
-  {% endwith %}
-{% endblock %}
