+++ tests/functional_tests/test_activity_validation.py
@@ -21,9 +21,11 @@ class ActivityValidationTest(AuthenticatedUserFunctionalTest):
 
     # They are redirected to their list of strava activities.
     # They click the link for the first activity presented.
-    section = self.browser.find_element(By.CLASS_NAME, 'content')
-    links = section.find_elements(By.TAG_NAME, 'a')
-    links[0].click()
+    datatable = self.wait_for_element(By.ID, 'datatable-activity')
+    datatable.find_elements(
+      By.XPATH, 
+      '//td[@data-dash-column="Title"]//a'
+    )[0].click()
 
     btn = self.wait_for_element(By.ID, 'save-activity')
     activity_url = self.browser.current_url
